refactor: create ResourceFetcherUtils namespace
